Attempt to improve GetFeatureInfo speed for habitats
The habitat layer contains many very complex features which are slow to
query for info.

In the mapfile reduce the info tolerance to 0 to
minamise the number of features returned by Postgres as the query for
features is quick but MapServer takes a while to process the returned
features. Also in the mapfile specify that we don't want geometery
returned in the response.

When we're dealing with a GetFeatureInfo request don't send the SLD_BODY
to MapServer as it's not required and therefore we don't want MapServer
considering the SLD.