Fix issues with paletted images being created by MapProxy
The global/image/paletted property defaults to true and results in 8bit
PNG tiles, we had an issue where we only have a few colours but they are
close to white and the quantizing was screwing them up. Setting paletted
to false results in 24bit PNG tiles and resolves the issue.