FMM-104: No GetFeatureInfo because TOLERANCE was set to 0, other fixes
Function makes use of BBOX internally for intersections and zero
TOLERANCE meant a BBOX with no area with resulted in null
intersections. Since moving the BBOX back out of the query would affect
speed, I've set the TOLERANCEUNITS to metres and TOLERANCE to 0.001 for
the sensitivity calculation layers.

Other fixes to functions to ensure returning all required fields and
that they are of the correct type; also ensured only returning correct
geometries in generalized sensitivity calculations.